@@ -3,15 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-scripts": "4.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "4.0.1",
     "uuid": "^8.3.2",
-    "values.js": "^2.0.0",
-    "web-vitals": "^1.1.0"
+    "values.js": "^2.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
