@@ -1,5 +1,5 @@
 import { Navbar, Sidebar, Footer } from "./components";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import HomeScreen from "./screen/HomeScreen";
 import AboutScreen from "./screen/AboutScreen";
 import ContactScreen from "./screen/ContactScreen";
@@ -8,16 +8,16 @@ import ErrorScreen from "./screen/ErrorScreen";
 
 function App() {
   return (
-    <Router className="App">
+    <Router className='App'>
       <Navbar />
       <Sidebar />
-      <Switch>
-        <Route path="/" exact component={HomeScreen} />
-        <Route path="/about" component={AboutScreen} />
-        <Route path="/contattaci" component={ContactScreen} />
-        <Route path="/cocktail/:id" component={SingleCocktailScreen} />
-        <Route path="*" component={ErrorScreen} />
-      </Switch>
+      <Routes>
+        <Route path='/' element={<HomeScreen />} />
+        <Route path='/about' element={<AboutScreen />} />
+        <Route path='/contattaci' element={<ContactScreen />} />
+        <Route path='/cocktail/:id' element={<SingleCocktailScreen />} />
+        <Route path='*' element={<ErrorScreen />} />
+      </Routes>
       <Footer />
     </Router>
   );
