@@ -1,5 +1,5 @@
 import { Navbar, Sidebar, Footer } from "./components";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import HomeScreen from "./screen/HomeScreen";
 import AboutScreen from "./screen/AboutScreen";
 import ContactScreen from "./screen/ContactScreen";
@@ -7,7 +7,7 @@ import SingleCocktailScreen from "./screen/SingleCocktailScreen";
 import ErrorScreen from "./screen/ErrorScreen";
 
 function App() {
-  return <div className="App">App</div>;
+  return <div className='App'>App</div>;
 }
 
 export default App;
