@@ -2,7 +2,7 @@ import React from "react";
 // import { Container, Stack, Box, Button } from "./styled";
 import { ReactComponent as Logo } from "../images/logo.svg";
 import { ReactComponent as CartIcon } from "../images/cart.svg";
-import { Link } from "@reach/router";
+import { Link } from "react-router-dom";
 const Header = () => {
   return (
     <>
