@@ -1,14 +1,21 @@
-import { Router } from "@reach/router";
 import HomePage from "./screen/home-screen";
 import Paginated from "./screen/paginated-screen";
 import Checkout from "./screen/checkout-screen";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 function App() {
   return (
-    <Router className='App'>
-      <HomePage path='/' />
-      <Checkout path='/checkout' />
-      <Paginated path='/photo/:page' />
+    <Router>
+      <Routes>
+        <Route path='/' element={<HomePage />} />
+        <Route path='/checkout/' element={<Checkout />} />
+        <Route
+          path='
+        /photo/:page
+        '
+          element={<Paginated />}
+        />
+      </Routes>
     </Router>
   );
 }
